refactor(api): migrate controller route to BC target profile
