+++ api/src/prescription/target-profile/application/admin-target-profile-controller.js
@@ -1,7 +1,9 @@
 import dayjs from 'dayjs';
 
+import * as queryParamsUtils from '../../../shared/infrastructure/utils/query-params-utils.js';
 import { escapeFileName } from '../../../shared/infrastructure/utils/request-response-utils.js';
 import { usecases } from '../domain/usecases/index.js';
+import * as organizationSerializer from '../infrastructure/serializers/jsonapi/organization-serializer.js';
 import * as targetProfileAttachOrganizationSerializer from '../infrastructure/serializers/jsonapi/target-profile-attach-organization-serializer.js';
 import * as targetProfileDetachOrganizationsSerializer from '../infrastructure/serializers/jsonapi/target-profile-detach-organizations-serializer.js';
 import * as targetProfileSerializer from '../infrastructure/serializers/jsonapi/target-profile-serializer.js';
@@ -98,6 +100,18 @@ const markTargetProfileAsSimplifiedAccess = async function (request, h) {
   return h.response(targetProfileSerializer.serialize(targetProfile));
 };
 
+const findPaginatedFilteredTargetProfileOrganizations = async function (request) {
+  const targetProfileId = request.params.id;
+  const options = queryParamsUtils.extractParameters(request.query);
+
+  const { models: organizations, pagination } = await usecases.findPaginatedFilteredOrganizationByTargetProfileId({
+    targetProfileId,
+    filter: options.filter,
+    page: options.page,
+  });
+  return organizationSerializer.serialize(organizations, pagination);
+};
+
 const targetProfileController = {
   outdateTargetProfile,
   markTargetProfileAsSimplifiedAccess,
@@ -107,6 +121,7 @@ const targetProfileController = {
   attachOrganizationsFromExistingTargetProfile,
   getContentAsJsonFile,
   getLearningContentAsPdf,
+  findPaginatedFilteredTargetProfileOrganizations,
 };
 
 export { targetProfileController };
