+++ api/tests/prescription/target-profile/unit/infrastructure/serializers/jsonapi/organization-serializer_test.js
@@ -1,7 +1,7 @@
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/organization-serializer.js';
-import { Organization } from '../../../../../src/organizational-entities/domain/models/Organization.js';
-import { Tag } from '../../../../../src/organizational-entities/domain/models/Tag.js';
-import { domainBuilder, expect } from '../../../../test-helper.js';
+import { Organization } from '../../../../../../../src/organizational-entities/domain/models/Organization.js';
+import { Tag } from '../../../../../../../src/organizational-entities/domain/models/Tag.js';
+import * as serializer from '../../../../../../../src/prescription/target-profile/infrastructure/serializers/jsonapi/organization-serializer.js';
+import { domainBuilder, expect } from '../../../../../../test-helper.js';
 
 describe('Unit | Serializer | organization-serializer', function () {
   describe('#serialize', function () {
