+++ api/tests/unit/application/target-profiles/index_test.js
@@ -217,148 +217,6 @@ describe('Unit | Application | Target Profiles | Routes', function () {
     });
   });
 
-  describe('GET /api/admin/target-profiles/{id}/organizations', function () {
-    const method = 'GET';
-    const url = '/api/admin/target-profiles/1/organizations';
-
-    context('when user has role "SUPER_ADMIN", "SUPPORT" or "METIER"', function () {
-      it('should return a response with an HTTP status code 200', async function () {
-        // given
-        sinon
-          .stub(securityPreHandlers, 'hasAtLeastOneAccessOf')
-          .withArgs([
-            securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-            securityPreHandlers.checkAdminMemberHasRoleSupport,
-            securityPreHandlers.checkAdminMemberHasRoleMetier,
-          ])
-          .callsFake(() => (request, h) => h.response(true));
-        sinon
-          .stub(targetProfileController, 'findPaginatedFilteredTargetProfileOrganizations')
-          .callsFake((request, h) => h.response('ok').code(200));
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        // when
-        const { statusCode } = await httpTestServer.request(method, url);
-
-        // then
-        expect(statusCode).to.equal(200);
-      });
-
-      context('when there is no filter nor pagination', function () {
-        it('should resolve with an HTTP status code 200', async function () {
-          // given
-          sinon.stub(securityPreHandlers, 'hasAtLeastOneAccessOf').returns(() => true);
-          sinon
-            .stub(targetProfileController, 'findPaginatedFilteredTargetProfileOrganizations')
-            .callsFake((request, h) => h.response('ok').code(200));
-          const httpTestServer = new HttpTestServer();
-          await httpTestServer.register(moduleUnderTest);
-
-          // when
-          const { statusCode } = await httpTestServer.request(method, url);
-
-          // then
-          expect(statusCode).to.equal(200);
-        });
-      });
-
-      context('when there are filters and pagination', function () {
-        it('should resolve with an HTTP status code 200', async function () {
-          // given
-          sinon.stub(securityPreHandlers, 'hasAtLeastOneAccessOf').returns(() => true);
-          sinon
-            .stub(targetProfileController, 'findPaginatedFilteredTargetProfileOrganizations')
-            .callsFake((request, h) => h.response('ok').code(200));
-          const httpTestServer = new HttpTestServer();
-          await httpTestServer.register(moduleUnderTest);
-
-          // when
-          const { statusCode } = await httpTestServer.request(
-            method,
-            `${url}?filter[name]=azerty&filter[type]=sco&filter[external-id]=abc&page[size]=10&page[number]=1`,
-          );
-
-          // then
-          expect(statusCode).to.equal(200);
-        });
-      });
-
-      context('when id is not an integer', function () {
-        it('should reject request with HTTP code 400', async function () {
-          // given
-          const httpTestServer = new HttpTestServer();
-          await httpTestServer.register(moduleUnderTest);
-
-          // when
-          const { statusCode } = await httpTestServer.request(
-            method,
-            '/api/admin/target-profiles/azerty/organizations',
-          );
-
-          // then
-          expect(statusCode).to.equal(400);
-        });
-      });
-
-      context('when page size is not an integer', function () {
-        it('should reject request with HTTP code 400', async function () {
-          // given
-          const httpTestServer = new HttpTestServer();
-          await httpTestServer.register(moduleUnderTest);
-
-          // when
-          const { statusCode } = await httpTestServer.request(method, `${url}?page[size]=azerty`);
-
-          // then
-          expect(statusCode).to.equal(400);
-        });
-      });
-
-      context('when page number is not an integer', function () {
-        it('should reject request with HTTP code 400', async function () {
-          // given
-          const httpTestServer = new HttpTestServer();
-          await httpTestServer.register(moduleUnderTest);
-
-          // when
-          const { statusCode } = await httpTestServer.request(method, `${url}?page[number]=azerty`);
-
-          // then
-          expect(statusCode).to.equal(400);
-        });
-      });
-    });
-
-    context('when user has role "CERTIF"', function () {
-      it('should return a response with an HTTP status code 403', async function () {
-        // given
-        sinon
-          .stub(securityPreHandlers, 'hasAtLeastOneAccessOf')
-          .withArgs([
-            securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-            securityPreHandlers.checkAdminMemberHasRoleSupport,
-            securityPreHandlers.checkAdminMemberHasRoleMetier,
-          ])
-          .callsFake(
-            () => (request, h) =>
-              h
-                .response({ errors: new Error('forbidden') })
-                .code(403)
-                .takeover(),
-          );
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        // when
-        const { statusCode } = await httpTestServer.request(method, url);
-
-        // then
-        expect(statusCode).to.equal(403);
-      });
-    });
-  });
-
   describe('POST /api/admin/target-profiles/{id}/badges', function () {
     const method = 'POST';
     const url = '/api/admin/target-profiles/123/badges';
